data-scraping/requirements.txt
Google-Images-Search==1.1.1
Pillow==6.2.2
